Adding sensitivity plots and updating few old ones

+++ build/costEvolution.py
@@ -1,8 +1,10 @@
+# %%
 import os
 import sys
 import numpy as np
 from matplotlib import pyplot as plt
 
+# %%
 plt.style.use('ggplot')
 plt.rcParams.update({
     "text.usetex": True,
@@ -16,56 +18,52 @@ plt.rcParams.update({
     "font.serif": ["Palatino"],
 })
 
+# %%
 startGen = 1
 endGen = 800
 steps = 1
+replayErrorThreshold = 20
+nPonitSmoothing= 10
 
-fileName = sys.argv[1]
-
+fileName = "results/00pct/logBest.txt"
 costWeights = np.array([800, 6, 0.3, 1.2, 175, 70, 10, 1000, 0.1])
 costLabels = [r"$\phi_1$", r"$\phi_9$", r"$\phi_{10}$", r"$\phi_2$", 
                 r"$\phi_3$", r"$\phi_4$", r"$\phi_6$", r"$\phi_8$",
                 r"$\phi_7$"]
 
 colors = plt.cm.tab10(np.linspace(0, 1, 11))
-# 1st generation is skipped as it has nan as cost value
+
+# %%
+
+# 1st generation is skipped as it has nan as the cost value
 logData = np.genfromtxt(fileName, delimiter=",", skip_header=1)
 logData = logData[range(startGen, endGen, steps), :]
 np.savetxt("filteredFile.csv", logData, delimiter=",")
 
-# Removing the costs file if it already exists
-if (os.path.exists("costsData.csv")):
-    os.remove("costsData.csv")
+if (os.path.exists("costsData.csv")): os.remove("costsData.csv")
 
-# Evaluating the costs
 numEvals = len(logData)
 os.system(f"./replay filteredFile.csv 1 {numEvals} false false")
-fileData = np.genfromtxt("costsData.csv", delimiter=",")
 
-# Sorting based on function Evaluations
+
+# %% Reads the costs data, sorts it, weighs it, filters it based on replay error
+# and smooths it using nPoint Moving Average
+
+fileData = np.genfromtxt("costsData.csv", delimiter=",")
 fileDataSorted = fileData[np.argsort(fileData[:, 0])]
 progress = 100*(1 - fileDataSorted[:,2])
 
-# Weighting the costs
 (rows, cols) = fileDataSorted.shape
 for i in range(cols-3):
     fileDataSorted[:,i+2] = costWeights[i]*fileDataSorted[:,i+2]
 
-# Removing gens with replay cost error
 replayCost = np.sum(fileDataSorted[:,2:-1], axis=1)
 simulationCost = logData[:,1]
-mask = []
-for i in range(replayCost.shape[0]):
-    if(abs(replayCost[i]-simulationCost[i])<20):
-        mask.append(True)
-    else:
-        mask.append(False)
+mask = [bool(replayCost[i] - simulationCost[i] < replayErrorThreshold) for i in range(len(replayCost))]
 
 fileDataSorted = fileDataSorted[mask,:]
 progress = progress[mask]
 
-# Smoothing the individual costs
-nPonitSmoothing= 10
 (rows, cols) = fileDataSorted.shape
 smoothedData = np.zeros((rows-nPonitSmoothing+1, cols))
 for i in range(cols):
@@ -74,49 +72,31 @@ for i in range(cols):
 fileDataSorted = smoothedData
 progress = np.convolve(progress, np.ones(nPonitSmoothing), "valid")/nPonitSmoothing
 
-## Relative contributions to total costs
 (rows, cols) = fileDataSorted.shape
 contributions = np.zeros((rows, cols-3))
 for i in range(cols-3):
     contributions[:,i] = np.divide(fileDataSorted[:,i+2], fileDataSorted[:,1])*100
 
-# Plotting the evaluation of costs
+# %% Plotting 
 gens = fileDataSorted[:,0]
-plt.figure(figsize=(5.5, 3), dpi=300)
-plt.plot(gens, fileDataSorted[:,1], "-", label=r"$\phi_{total}$", linewidth=1, color=colors[0], alpha=0.75)
 
-plt.plot(gens, fileDataSorted[:,10], label=costLabels[8], color=colors[9], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,6], label=costLabels[4], color=colors[5], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,7], label=costLabels[5], color=colors[6], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,2], label=costLabels[0], color=colors[1], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,3], label=costLabels[1], color=colors[2], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,4], label=costLabels[2], color=colors[3], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,9], label=costLabels[7], color=colors[8], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,8], label=costLabels[6], color=colors[7], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,5], label=costLabels[3], color=colors[4], alpha=0.75)
+plottingOrder = [8, 4, 5, 0, 1, 2, 7, 6 ,3]
+fig,axs = plt.subplots(2, 1, sharex=True, figsize=(6,4), dpi=300)
 
-plt.xlim(gens[0], gens[-1])
-plt.xlabel(r"$Generations$")
-plt.ylabel(r"$Value$")
-#plt.legend(loc="upper right")
+axs[0].plot(gens, fileDataSorted[:,1], "-", label=r"$\phi_{total}$", linewidth=1, color=colors[0], alpha=0.75)
+axs[1].plot(gens, progress, color="k", label=r"$100\times\alpha$", alpha=0.75)
+for i in plottingOrder:
+    axs[0].plot(gens, fileDataSorted[:,i+2], color=colors[i+1], label=costLabels[i], alpha=0.75)
+    axs[1].plot(gens, contributions[:,i], color=colors[i+1], label=costLabels[i], alpha=0.75)
+
+
+axs[1].set_xlim(gens[0], gens[-1])
+axs[0].set_ylabel("Value")
+axs[1].set_ylabel("Value")
+axs[1].set_xlabel("Generations")
+#axs[0].legend(loc="upper right")
+#axs[1].legend(loc="upper right")
+
 plt.savefig("CostsVsGen.png", format="png", transparent=False, bbox_inches='tight')
-
-plt.figure(figsize=(5.5, 3), dpi=300)
-plt.plot(gens, progress, color="k", label=r"$100\times\alpha$", alpha=0.75)
-plt.plot(gens, contributions[:,8], label=costLabels[8], color=colors[9], alpha=0.75)
-plt.plot(gens, contributions[:,4], label=costLabels[4], color=colors[5], alpha=0.75)
-plt.plot(gens, contributions[:,5], label=costLabels[5], color=colors[6], alpha=0.75)
-plt.plot(gens, contributions[:,0], label=costLabels[0], color=colors[1], alpha=0.75)
-plt.plot(gens, contributions[:,1], label=costLabels[1], color=colors[2], alpha=0.75)
-plt.plot(gens, contributions[:,2], label=costLabels[2], color=colors[3], alpha=0.75)
-plt.plot(gens, contributions[:,7], label=costLabels[7], color=colors[8], alpha=0.75)
-plt.plot(gens, contributions[:,6], label=costLabels[6], color=colors[7], alpha=0.75)
-plt.plot(gens, fileDataSorted[:,3], label=costLabels[3], color=colors[4], alpha=0.75)
-
-plt.xlim(gens[0], gens[-1])
-plt.xlabel(r"$Generations$")
-plt.ylabel(r"$Value$")
-#plt.legend(loc="upper right")
-plt.savefig("CostsContribVsGen.png", format="png", transparent=False, bbox_inches='tight')
-
 plt.show()
+# %%
